// graphql/generated/graphql.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  /** A date-time string at UTC, such as 2007-12-03T10:15:30Z, compliant with the `date-time` format outlined in section 5.6 of the RFC 3339 profile of the ISO 8601 standard for representation of dates and times using the Gregorian calendar. */
  DateTime: { input: any; output: any; }
  /** The `JSONObject` scalar type represents JSON objects as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSONObject: { input: any; output: any; }
  /** The `BigInt` scalar type represents non-fractional signed whole numeric values. */
  Long: { input: any; output: any; }
  /** Mongo object id scalar type */
  ObjectId: { input: any; output: any; }
};
// ...
export type CreateVariantPayload = {
  __typename?: 'CreateVariantPayload';
  /** The unique identifier for the incident. */
  incident_id: Scalars['Int']['output'];
  /** The unique report number associated with the incident. */
  report_number: Scalars['Int']['output'];
};
// ...
export type Entity = {
  __typename?: 'Entity';
  _id?: Maybe<Scalars['ObjectId']['output']>;
  created_at?: Maybe<Scalars['DateTime']['output']>;
  date_modified?: Maybe<Scalars['DateTime']['output']>;
  entity_id: Scalars['String']['output'];
  name: Scalars['String']['output'];
};
// ...
export type Incident = {
  __typename?: 'Incident';
  AllegedDeployerOfAISystem?: Maybe<Array<Maybe<Entity>>>;
  AllegedDeveloperOfAISystem?: Maybe<Array<Maybe<Entity>>>;
  AllegedHarmedOrNearlyHarmedParties?: Maybe<Array<Maybe<Entity>>>;
  _id?: Maybe<Scalars['ObjectId']['output']>;
  created_at?: Maybe<Scalars['DateTime']['output']>;
  date: Scalars['String']['output'];
  description?: Maybe<Scalars['String']['output']>;
  editor_dissimilar_incidents?: Maybe<Array<Maybe<Scalars['Int']['output']>>>;
  editor_notes: Scalars['String']['output'];
  editor_similar_incidents?: Maybe<Array<Maybe<Scalars['Int']['output']>>>;
  editors?: Maybe<Array<Maybe<User>>>;
  embedding?: Maybe<IncidentEmbedding>;
  epoch_date_modified?: Maybe<Scalars['Int']['output']>;
  flagged_dissimilar_incidents: Array<Maybe<Scalars['Int']['output']>>;
  implicated_systems?: Maybe<Array<Maybe<Entity>>>;
  incident_id: Scalars['Int']['output'];
  nlp_similar_incidents?: Maybe<Array<Maybe<IncidentNlp_Similar_Incident>>>;
  reports?: Maybe<Array<Maybe<Report>>>;
  title: Scalars['String']['output'];
  translations?: Maybe<Array<Maybe<IncidentTranslations>>>;
  tsne?: Maybe<IncidentTsne>;
};
// ...
export type IncidentEmbedding = {
  __typename?: 'IncidentEmbedding';
  from_reports?: Maybe<Array<Maybe<Scalars['Int']['output']>>>;
  vector?: Maybe<Array<Maybe<Scalars['Float']['output']>>>;
};
// ...
export type IncidentNlp_Similar_Incident = {
  __typename?: 'IncidentNlp_similar_incident';
  incident_id?: Maybe<Scalars['Int']['output']>;
  similarity?: Maybe<Scalars['Float']['output']>;
};
// ...
export type IncidentTranslations = {
  __typename?: 'IncidentTranslations';
  description?: Maybe<Scalars['String']['output']>;
  language?: Maybe<Scalars['String']['output']>;
  title?: Maybe<Scalars['String']['output']>;
};

export type IncidentTsne = {
  __typename?: 'IncidentTsne';
  x?: Maybe<Scalars['Float']['output']>;
  y?: Maybe<Scalars['Float']['output']>;
};
// ...
export type Report = {
  __typename?: 'Report';
  _id?: Maybe<Scalars['ObjectId']['output']>;
  authors: Array<Maybe<Scalars['String']['output']>>;
  cloudinary_id: Scalars['String']['output'];
  created_at?: Maybe<Scalars['DateTime']['output']>;
  date_downloaded: Scalars['DateTime']['output'];
  date_modified: Scalars['DateTime']['output'];
  date_published: Scalars['DateTime']['output'];
  date_submitted: Scalars['DateTime']['output'];
  description?: Maybe<Scalars['String']['output']>;
  editor_notes?: Maybe<Scalars['String']['output']>;
  embedding?: Maybe<ReportEmbedding>;
  epoch_date_downloaded: Scalars['Int']['output'];
  epoch_date_modified: Scalars['Int']['output'];
  epoch_date_published: Scalars['Int']['output'];
  epoch_date_submitted: Scalars['Int']['output'];
  flag?: Maybe<Scalars['Boolean']['output']>;
  image_url: Scalars['String']['output'];
  inputs_outputs?: Maybe<Array<Maybe<Scalars['String']['output']>>>;
  is_incident_report?: Maybe<Scalars['Boolean']['output']>;
  language: Scalars['String']['output'];
  plain_text: Scalars['String']['output'];
  quiet?: Maybe<Scalars['Boolean']['output']>;
  report_number: Scalars['Int']['output'];
  source_domain: Scalars['String']['output'];
  submitters: Array<Maybe<Scalars['String']['output']>>;
  tags: Array<Maybe<Scalars['String']['output']>>;
  text: Scalars['String']['output'];
  title: Scalars['String']['output'];
  translations?: Maybe<Array<Maybe<ReportTranslations>>>;
  url: Scalars['String']['output'];
  user?: Maybe<User>;
};


export type ReportTranslationsArgs = {
  languages: Array<InputMaybe<Scalars['String']['input']>>;
};

export type ReportEmbedding = {
  __typename?: 'ReportEmbedding';
  from_text_hash?: Maybe<Scalars['String']['output']>;
  vector?: Maybe<Array<Maybe<Scalars['Float']['output']>>>;
};
// ...
export type ReportTranslations = {
  __typename?: 'ReportTranslations';
  language: Scalars['String']['output'];
  plain_text?: Maybe<Scalars['String']['output']>;
  text?: Maybe<Scalars['String']['output']>;
  title?: Maybe<Scalars['String']['output']>;
};
// ...
export type User = {
  __typename?: 'User';
  _id?: Maybe<Scalars['ObjectId']['output']>;
  adminData?: Maybe<UserAdminDatum>;
  first_name?: Maybe<Scalars['String']['output']>;
  last_name?: Maybe<Scalars['String']['output']>;
  roles: Array<Maybe<Scalars['String']['output']>>;
  userId: Scalars['String']['output'];
};

export type UserAdminDatum = {
  __typename?: 'UserAdminDatum';
  creationDate: Scalars['DateTime']['output'];
  disabled: Scalars['Boolean']['output'];
  email: Scalars['String']['output'];
  lastAuthenticationDate?: Maybe<Scalars['DateTime']['output']>;
};